requirements.txt
# Core dependencies for Pythonium MCP Server
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0
loguru>=0.7.0
asyncio-mqtt>=0.16.0
websockets>=11.0.0
aiohttp>=3.8.0
httpx>=0.24.0
jsonschema>=4.0.0
packaging>=21.0
psutil>=5.9.0

# Performance optimizations
orjson>=3.9.0  # High-performance JSON serialization
aiofiles>=23.0.0  # Async file I/O operations

# Optional dependencies for enhanced features
watchdog>=3.0.0  # File system monitoring for configuration hot-reload

# Development dependencies (install with: pip install -r requirements.txt -r requirements-dev.txt)
# See requirements-dev.txt for development-specific packages
